Add files via upload
A bag consists of zero or more objects of the same type. Each object can be described by its color and weight. Design a Java program to create a new object. This can be done in two ways. If the user provides information about color and/or weight of the object to be created then this information will be used to create the object otherwise the object will be created using default values for these attributes. Provide a facility to keep track of total number of objects and total weight of object from a bag. Use static variable and method, array of object/reference, and constructor overloading.