Add files via upload
Design the 'employee' class with a name, id, DOB,  mail_id, and mobile_no as instance variables. Inherit the classes: Programmer, Team Lead, Assistant Project Manager, and Project Manager from the employee class. Add Basic Pay (BP) as the member of all the
inherited classes with 97% of BP as DA, 10 % of BP as HRA, 12% of BP as PF, and 0.1% of BP for staff
club fund. Generate pay slips for the employees with their gross and net salary.