Add files via upload
Design a base class 'Shape' with double type value and methods compute_area() for calculating the area of shape. Derive three classes: triangle, rectangle, and circle. Make compute_area() as an abstract method and redefine this method in the derived class to suit their requirements. Write a program that accepts dimensions of triangle/rectangle/circle and displays the calculated area. Implement dynamic binding for a given case study.