Add files via upload
Implement a program to handle Arithmetic Exception, and Input Mismatch Exception. The user enters two numbers Num1 and Num2. The division of Num1 and Num2 is displayed. If Num1 and Num2 are not integers, the program would throw an Input Mismatch Exception. If Num2 were zero, the program would throw an Arithmetic Exception. Display the exception.